requirements.txt
Django>=4.1,<4.1.5
django-cors-headers>=3.11.0,<3.14.0
django-rest-swagger>=2.2.0,<2.3.0
djangorestframework>=3.13.0,<3.15.0
drf-yasg==1.21.4
gunicorn>=20.0.4,<20.2.0
psycopg2-binary==2.9.5
python-dotenv==0.21.0
redis==4.5.1
django-redis==5.2.0
celery==5.2.7
django-celery-beat==2.5.0